requirements.txt
# 多AI協作開發平台依賴套件

# 核心框架
flask>=2.3.0
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP客戶端和API調用
requests>=2.31.0
aiohttp>=3.8.0
httpx>=0.25.0

# AI API客戶端
openai>=1.3.0
anthropic>=0.18.0

# 資料處理
pyyaml>=6.0
python-dotenv>=1.0.0

# 檔案監控
watchdog>=3.0.0

# 日誌和調試
logging>=0.4.9.6

# 異步處理
asyncio
aiofiles>=23.0.0

# JSON處理和驗證
jsonschema>=4.20.0

# 日期時間處理
python-dateutil>=2.8.0

# Web前端支援
jinja2>=3.1.0
werkzeug>=2.3.0

# 測試工具（開發環境）
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# 程式碼品質工具（開發環境）
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# 安全性
cryptography>=41.0.0

# 效能監控（可選）
psutil>=5.9.0

# 文件生成（可選）
markdown>=3.5.0